// src/app/interfaces/IFavorites.ts
export interface IFavorites {
    email: string;
    breedSpeciesName: string;
    breedSpeciesPicture: string;
    height: string;
    weight: string;
    lifespan: string;
    diet: string;
    dangerousFood: string;
    goodFood: string;
    about: string;
    groomingAndHealth: string,
    careAndHabitat: string,
    personality: string;
    category: string;
}